Unmask the PHP 5.5 target USE flag
PHP 5.5 has reached end of life and will be removed in 90 days from
the Gentoo main portage tree.

Allow PHP 5.5 to be installed until all systems are upgraded to PHP >
5.5.

The dev-lang/php and virtual/httpd-php packages still have to be
unmasked manually, in order to raise awareness of the upgrade.